+++ src/types/generic/normal.rs
@@ -3,6 +3,7 @@
 #[cfg(feature = "alloc")]
 use alloc::string::String;
 
+use crate::components::AuthorityComponents;
 use crate::parser::trusted as trusted_parser;
 #[cfg(feature = "alloc")]
 use crate::raw;
@@ -328,6 +329,37 @@ impl<S: Spec> RiStr<S> {
     #[must_use]
     pub fn fragment(&self) -> Option<&RiFragmentStr<S>> {
         AsRef::<RiReferenceStr<S>>::as_ref(self).fragment()
+    }
+
+    /// Returns the authority components.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use iri_string::validate::Error;
+    /// use iri_string::types::IriStr;
+    ///
+    /// let iri = IriStr::new("http://user:pass@example.com:8080/pathpath?queryquery")?;
+    /// let authority = iri.authority_components()
+    ///     .expect("authority is available");
+    /// assert_eq!(authority.userinfo(), Some("user:pass"));
+    /// assert_eq!(authority.host(), "example.com");
+    /// assert_eq!(authority.port(), Some("8080"));
+    /// # Ok::<_, Error>(())
+    /// ```
+    ///
+    /// ```
+    /// # use iri_string::validate::Error;
+    /// use iri_string::types::IriStr;
+    ///
+    /// let iri = IriStr::new("urn:uuid:10db315b-fcd1-4428-aca8-15babc9a2da2")?;
+    /// assert_eq!(iri.authority_str(), None);
+    /// # Ok::<_, Error>(())
+    /// ```
+    #[inline]
+    #[must_use]
+    pub fn authority_components(&self) -> Option<AuthorityComponents<'_>> {
+        AuthorityComponents::from_iri(self.as_ref())
     }
 }