+++ src/types/generic/reference.rs
@@ -5,6 +5,7 @@ use core::convert::TryFrom;
 #[cfg(feature = "alloc")]
 use alloc::{borrow::Cow, string::String};
 
+use crate::components::AuthorityComponents;
 use crate::parser::trusted as trusted_parser;
 #[cfg(feature = "alloc")]
 use crate::raw;
@@ -388,6 +389,37 @@ impl<S: Spec> RiReferenceStr<S> {
             // returned string is substring of the source IRI.
             RiFragmentStr::new_maybe_unchecked(fragment)
         })
+    }
+
+    /// Returns the authority components.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use iri_string::validate::Error;
+    /// use iri_string::types::IriReferenceStr;
+    ///
+    /// let iri = IriReferenceStr::new("http://user:pass@example.com:8080/pathpath?queryquery")?;
+    /// let authority = iri.authority_components()
+    ///     .expect("authority is available");
+    /// assert_eq!(authority.userinfo(), Some("user:pass"));
+    /// assert_eq!(authority.host(), "example.com");
+    /// assert_eq!(authority.port(), Some("8080"));
+    /// # Ok::<_, Error>(())
+    /// ```
+    ///
+    /// ```
+    /// # use iri_string::validate::Error;
+    /// use iri_string::types::IriReferenceStr;
+    ///
+    /// let iri = IriReferenceStr::new("foo//bar:baz")?;
+    /// assert_eq!(iri.authority_str(), None);
+    /// # Ok::<_, Error>(())
+    /// ```
+    #[inline]
+    #[must_use]
+    pub fn authority_components(&self) -> Option<AuthorityComponents<'_>> {
+        AuthorityComponents::from_iri(self)
     }
 }